Support raw identifiers
