+++ src/ser/mod.rs
@@ -1,9 +1,14 @@
 use serde::{ser, Deserialize, Serialize};
 use std::io;
 
-use crate::error::{Error, Result};
-use crate::extensions::Extensions;
+use crate::{
+    error::{Error, Result},
+    extensions::Extensions,
+    parse::{is_ident_first_char, is_ident_other_char},
+};
 
+#[cfg(test)]
+mod tests;
 mod value;
 
 /// Serializes `value` into `writer`
@@ -335,6 +340,15 @@ impl<W: io::Write> Serializer<W> {
         self.output.write_all(b"\"")?;
         Ok(())
     }
+
+    fn write_identifier(&mut self, name: &str) -> io::Result<()> {
+        let mut bytes = name.as_bytes().iter().cloned();
+        if !bytes.next().map_or(false, is_ident_first_char) || !bytes.all(is_ident_other_char) {
+            self.output.write_all(b"r#")?;
+        }
+        self.output.write_all(name.as_bytes())?;
+        Ok(())
+    }
 }
 
 impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
@@ -466,7 +480,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
 
     fn serialize_unit_struct(self, name: &'static str) -> Result<()> {
         if self.struct_names {
-            self.output.write_all(name.as_bytes())?;
+            self.write_identifier(name)?;
 
             Ok(())
         } else {
@@ -475,7 +489,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
     }
 
     fn serialize_unit_variant(self, _: &'static str, _: u32, variant: &'static str) -> Result<()> {
-        self.output.write_all(variant.as_bytes())?;
+        self.write_identifier(variant)?;
 
         Ok(())
     }
@@ -485,7 +499,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
         T: ?Sized + Serialize,
     {
         if self.struct_names {
-            self.output.write_all(name.as_bytes())?;
+            self.write_identifier(name)?;
         }
 
         self.output.write_all(b"(")?;
@@ -504,7 +518,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
     where
         T: ?Sized + Serialize,
     {
-        self.output.write_all(variant.as_bytes())?;
+        self.write_identifier(variant)?;
         self.output.write_all(b"(")?;
 
         value.serialize(&mut *self)?;
@@ -553,7 +567,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
         len: usize,
     ) -> Result<Self::SerializeTupleStruct> {
         if self.struct_names {
-            self.output.write_all(name.as_bytes())?;
+            self.write_identifier(name)?;
         }
 
         self.serialize_tuple(len)
@@ -566,7 +580,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
         variant: &'static str,
         len: usize,
     ) -> Result<Self::SerializeTupleVariant> {
-        self.output.write_all(variant.as_bytes())?;
+        self.write_identifier(variant)?;
         self.output.write_all(b"(")?;
 
         if self.separate_tuple_members() {
@@ -598,7 +612,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
 
     fn serialize_struct(self, name: &'static str, len: usize) -> Result<Self::SerializeStruct> {
         if self.struct_names {
-            self.output.write_all(name.as_bytes())?;
+            self.write_identifier(name)?;
         }
         self.output.write_all(b"(")?;
 
@@ -618,7 +632,7 @@ impl<'a, W: io::Write> ser::Serializer for &'a mut Serializer<W> {
         variant: &'static str,
         len: usize,
     ) -> Result<Self::SerializeStructVariant> {
-        self.output.write_all(variant.as_bytes())?;
+        self.write_identifier(variant)?;
         self.output.write_all(b"(")?;
 
         self.is_empty = Some(len == 0);
@@ -961,129 +975,5 @@ impl<'a, W: io::Write> ser::SerializeStructVariant for Compound<'a, W> {
 
     fn end(self) -> Result<()> {
         ser::SerializeStruct::end(self)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[derive(Serialize)]
-    struct EmptyStruct1;
-
-    #[derive(Serialize)]
-    struct EmptyStruct2 {}
-
-    #[derive(Serialize)]
-    struct MyStruct {
-        x: f32,
-        y: f32,
-    }
-
-    #[derive(Serialize)]
-    enum MyEnum {
-        A,
-        B(bool),
-        C(bool, f32),
-        D { a: i32, b: i32 },
-    }
-
-    #[test]
-    fn test_empty_struct() {
-        assert_eq!(to_string(&EmptyStruct1).unwrap(), "()");
-        assert_eq!(to_string(&EmptyStruct2 {}).unwrap(), "()");
-    }
-
-    #[test]
-    fn test_struct() {
-        let my_struct = MyStruct { x: 4.0, y: 7.0 };
-
-        assert_eq!(to_string(&my_struct).unwrap(), "(x:4,y:7)");
-
-        #[derive(Serialize)]
-        struct NewType(i32);
-
-        assert_eq!(to_string(&NewType(42)).unwrap(), "(42)");
-
-        #[derive(Serialize)]
-        struct TupleStruct(f32, f32);
-
-        assert_eq!(to_string(&TupleStruct(2.0, 5.0)).unwrap(), "(2,5)");
-    }
-
-    #[test]
-    fn test_option() {
-        assert_eq!(to_string(&Some(1u8)).unwrap(), "Some(1)");
-        assert_eq!(to_string(&None::<u8>).unwrap(), "None");
-    }
-
-    #[test]
-    fn test_enum() {
-        assert_eq!(to_string(&MyEnum::A).unwrap(), "A");
-        assert_eq!(to_string(&MyEnum::B(true)).unwrap(), "B(true)");
-        assert_eq!(to_string(&MyEnum::C(true, 3.5)).unwrap(), "C(true,3.5)");
-        assert_eq!(to_string(&MyEnum::D { a: 2, b: 3 }).unwrap(), "D(a:2,b:3)");
-    }
-
-    #[test]
-    fn test_array() {
-        let empty: [i32; 0] = [];
-        assert_eq!(to_string(&empty).unwrap(), "()");
-        let empty_ref: &[i32] = &empty;
-        assert_eq!(to_string(&empty_ref).unwrap(), "[]");
-
-        assert_eq!(to_string(&[2, 3, 4i32]).unwrap(), "(2,3,4)");
-        assert_eq!(to_string(&(&[2, 3, 4i32] as &[i32])).unwrap(), "[2,3,4]");
-    }
-
-    #[test]
-    fn test_map() {
-        use std::collections::HashMap;
-
-        let mut map = HashMap::new();
-        map.insert((true, false), 4);
-        map.insert((false, false), 123);
-
-        let s = to_string(&map).unwrap();
-        s.starts_with("{");
-        s.contains("(true,false):4");
-        s.contains("(false,false):123");
-        s.ends_with("}");
-    }
-
-    #[test]
-    fn test_string() {
-        assert_eq!(to_string(&"Some string").unwrap(), "\"Some string\"");
-    }
-
-    #[test]
-    fn test_char() {
-        assert_eq!(to_string(&'c').unwrap(), "'c'");
-    }
-
-    #[test]
-    fn test_escape() {
-        assert_eq!(to_string(&r#""Quoted""#).unwrap(), r#""\"Quoted\"""#);
-    }
-
-    #[test]
-    fn test_byte_stream() {
-        use serde_bytes;
-
-        let small: [u8; 16] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15];
-        assert_eq!(
-            to_string(&small).unwrap(),
-            "(0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15)"
-        );
-
-        let large = vec![255u8; 64];
-        let large = serde_bytes::Bytes::new(&large);
-        assert_eq!(
-            to_string(&large).unwrap(),
-            concat!(
-                "\"/////////////////////////////////////////",
-                "////////////////////////////////////////////w==\""
-            )
-        );
     }
 }
